requirements.txt
diff-match-patch==20121119
Django==1.9
django-crispy-forms==1.7.0
django-formtools==1.0
django-import-export==0.6.1
django-ranged-response==0.2.0
django-reversion==2.0.12
django-simple-captcha==0.4.6
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
jdcal==1.3
MySQL-python==1.2.5
odfpy==1.3.6
olefile==0.44
openpyxl==2.4.9
Pillow==4.3.0
PyYAML==3.12
six==1.10.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
